requirements/quality.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    make upgrade
#
amqp==5.3.1
    # via
    #   -r requirements/test.txt
    #   kombu
asgiref==3.8.1
    # via
    #   -r requirements/test.txt
    #   django
astroid==3.3.8
    # via
    #   pylint
    #   pylint-celery
attrs==24.3.0
    # via
    #   -r requirements/test.txt
    #   openedx-events
backports-tarfile==1.2.0
    # via jaraco-context
billiard==4.2.1
    # via
    #   -r requirements/test.txt
    #   celery
celery==5.4.0
    # via -r requirements/test.txt
certifi==2024.12.14
    # via
    #   -r requirements/test.txt
    #   requests
cffi==1.17.1
    # via
    #   -r requirements/test.txt
    #   cryptography
    #   pynacl
charset-normalizer==3.4.1
    # via
    #   -r requirements/test.txt
    #   requests
click==8.1.8
    # via
    #   -r requirements/test.txt
    #   celery
    #   click-didyoumean
    #   click-log
    #   click-plugins
    #   click-repl
    #   code-annotations
    #   edx-django-utils
    #   edx-lint
click-didyoumean==0.3.1
    # via
    #   -r requirements/test.txt
    #   celery
click-log==0.4.0
    # via edx-lint
click-plugins==1.1.1
    # via
    #   -r requirements/test.txt
    #   celery
click-repl==0.3.0
    # via
    #   -r requirements/test.txt
    #   celery
code-annotations==2.2.0
    # via
    #   -r requirements/test.txt
    #   edx-lint
coverage[toml]==7.6.10
    # via
    #   -r requirements/test.txt
    #   pytest-cov
cryptography==44.0.0
    # via secretstorage
ddt==1.7.2
    # via -r requirements/test.txt
dill==0.3.9
    # via pylint
django==4.2.18
    # via
    #   -c https://raw.githubusercontent.com/openedx/edx-lint/master/edx_lint/files/common_constraints.txt
    #   -r requirements/test.txt
    #   django-crum
    #   django-waffle
    #   edx-django-utils
    #   openedx-events
django-crum==0.7.9
    # via
    #   -r requirements/test.txt
    #   edx-django-utils
django-waffle==4.2.0
    # via
    #   -r requirements/test.txt
    #   edx-django-utils
dnspython==2.7.0
    # via
    #   -r requirements/test.txt
    #   pymongo
docutils==0.21.2
    # via readme-renderer
edx-ccx-keys==2.0.2
    # via
    #   -r requirements/test.txt
    #   openedx-events
edx-django-utils==7.1.0
    # via
    #   -r requirements/test.txt
    #   openedx-events
edx-lint==5.4.1
    # via -r requirements/quality.in
edx-opaque-keys[django]==2.11.0
    # via
    #   -r requirements/test.txt
    #   edx-ccx-keys
    #   openedx-events
fastavro==1.10.0
    # via
    #   -r requirements/test.txt
    #   openedx-events
idna==3.10
    # via
    #   -r requirements/test.txt
    #   requests
importlib-metadata==8.5.0
    # via keyring
iniconfig==2.0.0
    # via
    #   -r requirements/test.txt
    #   pytest
isort==5.13.2
    # via
    #   -r requirements/quality.in
    #   pylint
jaraco-classes==3.4.0
    # via keyring
jaraco-context==6.0.1
    # via keyring
jaraco-functools==4.1.0
    # via keyring
jeepney==0.8.0
    # via
    #   keyring
    #   secretstorage
jinja2==3.1.5
    # via
    #   -r requirements/test.txt
    #   code-annotations
keyring==25.6.0
    # via twine
kombu==5.4.2
    # via
    #   -r requirements/test.txt
    #   celery
markdown-it-py==3.0.0
    # via rich
markupsafe==3.0.2
    # via
    #   -r requirements/test.txt
    #   jinja2
mccabe==0.7.0
    # via pylint
mdurl==0.1.2
    # via markdown-it-py
more-itertools==10.6.0
    # via
    #   jaraco-classes
    #   jaraco-functools
newrelic==10.4.0
    # via
    #   -r requirements/test.txt
    #   edx-django-utils
nh3==0.2.20
    # via readme-renderer
openedx-events==9.15.2
    # via -r requirements/test.txt
packaging==24.2
    # via
    #   -r requirements/test.txt
    #   pytest
    #   twine
pbr==6.1.0
    # via
    #   -r requirements/test.txt
    #   stevedore
pkginfo==1.12.0
    # via twine
platformdirs==4.3.6
    # via pylint
pluggy==1.5.0
    # via
    #   -r requirements/test.txt
    #   pytest
prompt-toolkit==3.0.49
    # via
    #   -r requirements/test.txt
    #   click-repl
psutil==6.1.1
    # via
    #   -r requirements/test.txt
    #   edx-django-utils
pycodestyle==2.12.1
    # via -r requirements/quality.in
pycparser==2.22
    # via
    #   -r requirements/test.txt
    #   cffi
pydocstyle==6.3.0
    # via -r requirements/quality.in
pygments==2.19.1
    # via
    #   readme-renderer
    #   rich
pylint==3.3.3
    # via
    #   edx-lint
    #   pylint-celery
    #   pylint-django
    #   pylint-plugin-utils
pylint-celery==0.3
    # via edx-lint
pylint-django==2.6.1
    # via edx-lint
pylint-plugin-utils==0.8.2
    # via
    #   pylint-celery
    #   pylint-django
pymongo==4.10.1
    # via
    #   -r requirements/test.txt
    #   edx-opaque-keys
pynacl==1.5.0
    # via
    #   -r requirements/test.txt
    #   edx-django-utils
pytest==8.3.4
    # via
    #   -r requirements/test.txt
    #   pytest-cov
    #   pytest-django
pytest-cov==6.0.0
    # via -r requirements/test.txt
pytest-django==4.9.0
    # via -r requirements/test.txt
python-dateutil==2.9.0.post0
    # via
    #   -r requirements/test.txt
    #   celery
python-slugify==8.0.4
    # via
    #   -r requirements/test.txt
    #   code-annotations
pyyaml==6.0.2
    # via
    #   -r requirements/test.txt
    #   code-annotations
readme-renderer==44.0
    # via twine
requests==2.32.3
    # via
    #   -r requirements/test.txt
    #   requests-toolbelt
    #   twine
requests-toolbelt==1.0.0
    # via twine
rfc3986==2.0.0
    # via twine
rich==13.9.4
    # via twine
secretstorage==3.3.3
    # via keyring
six==1.17.0
    # via
    #   -r requirements/test.txt
    #   edx-ccx-keys
    #   edx-lint
    #   python-dateutil
snowballstemmer==2.2.0
    # via pydocstyle
sqlparse==0.5.3
    # via
    #   -r requirements/test.txt
    #   django
stevedore==5.4.0
    # via
    #   -r requirements/test.txt
    #   code-annotations
    #   edx-django-utils
    #   edx-opaque-keys
text-unidecode==1.3
    # via
    #   -r requirements/test.txt
    #   python-slugify
tomlkit==0.13.2
    # via pylint
twine==6.0.1
    # via -r requirements/quality.in
typing-extensions==4.12.2
    # via
    #   -r requirements/test.txt
    #   edx-opaque-keys
tzdata==2024.2
    # via
    #   -r requirements/test.txt
    #   celery
    #   kombu
urllib3==2.2.3
    # via
    #   -c https://raw.githubusercontent.com/openedx/edx-lint/master/edx_lint/files/common_constraints.txt
    #   -r requirements/test.txt
    #   requests
    #   twine
vine==5.1.0
    # via
    #   -r requirements/test.txt
    #   amqp
    #   celery
    #   kombu
wcwidth==0.2.13
    # via
    #   -r requirements/test.txt
    #   prompt-toolkit
zipp==3.21.0
    # via importlib-metadata
